Inline CallbackSignatureShouldMatchMockedMethod test data into tests
