+++ Source/Moq.Analyzers.Test/CallbackSignatureShouldMatchMockedMethodAnalyzerTests.cs
@@ -1,46 +1,38 @@
 ﻿using System.IO;
 using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.Diagnostics;
-using TestHelper;
 using Xunit;
 
 namespace Moq.Analyzers.Test;
 
-// TODO: These tests should be broken down further
-public class CallbackSignatureShouldMatchMockedMethodAnalyzerTests : DiagnosticVerifier
+public class CallbackSignatureShouldMatchMockedMethodAnalyzerTests : CallbackSignatureShouldMatchMockedMethodBase
 {
     [Fact]
     public Task ShouldPassWhenCorrectSetupAndReturns()
     {
-        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.MyGoodSetupAndReturns.cs")));
+        return Verify(VerifyCSharpDiagnostic(GoodSetupAndCallback));
     }
 
     [Fact]
     public Task ShouldFailWhenIncorrectCallbacks()
     {
-        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.TestBadCallbacks.cs")));
+        return Verify(VerifyCSharpDiagnostic(BadCallbacks));
     }
 
     [Fact]
     public Task ShouldPassWhenCorrectSetupAndCallbacks()
     {
-        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.TestGoodSetupAndCallback.cs")));
+        return Verify(VerifyCSharpDiagnostic(GoodSetupAndCallback));
     }
 
     [Fact]
     public Task ShouldPassWhenCorrectSetupAndParameterlessCallbacks()
     {
-        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.TestGoodSetupAndParameterlessCallback.cs")));
+        return Verify(VerifyCSharpDiagnostic(GoodSetupAndParameterlessCallback));
     }
 
     [Fact]
     public Task ShouldPassWhenCorrectSetupAndReturnsAndCallbacks()
     {
-        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/CallbackSignatureShouldMatchMockedMethod.TestGoodSetupAndReturnsAndCallback.cs")));
-    }
-
-    protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
-    {
-        return new CallbackSignatureShouldMatchMockedMethodAnalyzer();
+        return Verify(VerifyCSharpDiagnostic(GoodSetupAndReturnsAndCallback));
     }
 }
