--- Source/Moq.Analyzers.Test/Data/CallbackSignatureShouldMatchMockedMethod.MyGoodSetupAndReturns.cs
@@ -1,25 +1,0 @@
-using System;
-using System.Collections.Generic;
-using Moq;
-
-namespace CallbackSignatureShouldMatchMockedMethod.MyGoodSetupAndReturns;
-
-internal interface IFoo
-{
-    int Do(string s);
-
-    int Do(int i, string s, DateTime dt);
-
-    int Do(List<string> l);
-}
-
-internal class MyUnitTests
-{
-    private void MyGoodSetupAndReturns()
-    {
-        var mock = new Mock<IFoo>();
-        mock.Setup(x => x.Do(It.IsAny<string>())).Returns((string s) => { return 0; });
-        mock.Setup(x => x.Do(It.IsAny<int>(), It.IsAny<string>(), It.IsAny<DateTime>())).Returns((int i, string s, DateTime dt) => { return 0; });
-        mock.Setup(x => x.Do(It.IsAny<List<string>>())).Returns((List<string> l) => { return 0; });
-    }
-}
